move AutoString/FileNode/SourceMap out of stdafx
