+++ chakracore-cxx/bin/ch/stdafx.h
@@ -151,192 +151,3 @@ do { \
 #include "MessageQueue.h"
 #include "RuntimeThreadData.h"
 #include "WScriptJsrt.h"
-
-class AutoString
-{
-    size_t length;
-    char* data;
-    char16_t* data_wide;
-    JsErrorCode errorCode;
-    bool dontFree;
-public:
-    AutoString():length(0), data(nullptr),
-        data_wide(nullptr), errorCode(JsNoError), dontFree(false)
-    { }
-
-    AutoString(AutoString &autoString):length(autoString.length),
-        data(autoString.data), data_wide(autoString.data_wide),
-        errorCode(JsNoError), dontFree(false)
-    {
-        autoString.dontFree = true; // take over the ownership
-    }
-
-    AutoString(JsValueRef value):length(0), data(nullptr),
-        data_wide(nullptr), errorCode(JsNoError), dontFree(false)
-    {
-        Initialize(value);
-    }
-
-    JsErrorCode Initialize(JsValueRef value)
-    {
-        errorCode = JsNoError;
-        JsValueRef strValue;
-        JsValueType type;
-        ChakraRTInterface::JsGetValueType(value, &type);
-        if (type != JsString)
-        {
-            errorCode = ChakraRTInterface::JsConvertValueToString(value, &strValue);
-        }
-        else
-        {
-            strValue = value;
-        }
-        size_t length = 0;
-        if (errorCode == JsNoError)
-        {
-            errorCode = ChakraRTInterface::JsCopyString(strValue, nullptr, 0, &length);
-            if (errorCode == JsNoError)
-            {
-                data = (char*)malloc((length + 1) * sizeof(char));
-                size_t writtenLength = 0;
-                errorCode = ChakraRTInterface::JsCopyString(strValue, data, length, &writtenLength);
-                if (errorCode == JsNoError)
-                {
-                    AssertMsg(length == writtenLength, "Inconsistent length in utf8 encoding");
-                }
-            }
-        }
-        if (errorCode == JsNoError)
-        {
-            *(data + length) = char(0);
-            this->length = length;
-        }
-        return errorCode;
-    }
-
-    void MakePersistent()
-    {
-        dontFree = true;
-    }
-
-    const char * GetString()
-    {
-        return data;
-    }
-
-    char16_t* GetWideString(charcount_t* destCount = nullptr)
-    {
-        if(data_wide || !data)
-        {
-            return data_wide;
-        }
-        charcount_t tempDestCount;
-        utf8::NarrowStringToWide<utf8::malloc_allocator>(data, length, &data_wide, &tempDestCount);
-
-        if (destCount)
-        {
-            *destCount = tempDestCount;
-        }
-        return data_wide;
-    }
-
-    bool HasError()
-    {
-        return errorCode != JsNoError;
-    }
-
-    JsErrorCode GetError()
-    {
-        return errorCode;
-    }
-
-    size_t GetLength()
-    {
-        return length;
-    }
-
-    ~AutoString()
-    {
-        // we need persistent source string
-        // for externalArrayBuffer source
-        // externalArrayBuffer finalize should
-        // free this memory
-        if (!dontFree && data != nullptr)
-        {
-            free(data);
-            data = nullptr;
-        }
-
-        // Free this anyway.
-        if (data_wide != nullptr)
-        {
-            free(data_wide);
-            data_wide = nullptr;
-        }
-    }
-
-    char* operator*() { return data; }
-};
-
-struct FileNode
-{
-    AutoString data;
-    AutoString path;
-    FileNode * next;
-    FileNode(AutoString &path_, AutoString &data_):
-        path(path_), data(data_), next(nullptr) {
-        path_.MakePersistent();
-        data_.MakePersistent();
-    }
-};
-
-class SourceMap
-{
-    static FileNode * root;
-public:
-    static void Add(AutoString &path, AutoString &data)
-    {
-        // SourceMap lifetime == process lifetime
-        FileNode * node = new FileNode(path, data);
-        if (root != nullptr)
-        {
-            node->next = root;
-        }
-        root = node;
-    }
-
-    static bool Find(AutoString &path, AutoString ** out)
-    {
-        return Find(path.GetString(), path.GetLength(), out);
-    }
-
-    static bool Find(const char * path, size_t pathLength, AutoString ** out)
-    {
-        FileNode * node = root;
-        while(node != nullptr)
-        {
-            if (strncmp(node->path.GetString(), path, pathLength) == 0)
-            {
-                *out = &(node->data);
-                return true;
-            }
-            node = node->next;
-        }
-        return false;
-    }
-
-    static bool Find(const std::filesystem::path &path, AutoString **out)
-    {
-        FileNode *node = root;
-        while (node != nullptr)
-        {
-            if (node->path.GetString() == path.native())
-            {
-                *out = &(node->data);
-                return true;
-            }
-            node = node->next;
-        }
-        return false;
-    }
-};
